+++ tests/models/test_emails.py
@@ -117,7 +117,7 @@ class TestSimpleSend:
     def test_invalid(self, server_client):
         with pytest.raises(TypeError) as exc:
             server_client.emails.send(message=object())
-        assert str(exc.value) == 'message should be either Email or MIMEText instance'
+        assert str(exc.value) == 'message should be either Email or MIMEText or MIMEMultipart instance'
 
     def test_message_and_kwargs(self, server_client, email):
         with pytest.raises(AssertionError) as exc: